requirements.txt
Flask==2.3.3
pandas==2.2.2
unidecode==1.3.8
cloudinary==1.40.0
python-dotenv==1.0.1
gunicorn==22.0.0
openpyxl==3.1.2
requests==2.31.0
numpy==1.26.4
msal==1.20.0
Flask-Login==0.6.3
Werkzeug>=2.3.7
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
